# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moneykeeper",
    "version": "0.1.0-foundation",
    "description": "Moneykeeper context-first contract mounted by the default runtime at unversioned paths."
  },
  "paths": {
    "/portfolio-accounts": {
      "get": {"operationId":"listPortfolioAccounts","summary":"List portfolio accounts","description":"Returns active and archived manual portfolio accounts for the authenticated owner.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Portfolio accounts"},"401":{"$ref":"#/components/responses/Unauthorized"},"500":{"$ref":"#/components/responses/InternalServerError"}}},
      "post": {"operationId":"openPortfolioAccount","summary":"Open a portfolio account","description":"Creates a manual portfolio account at version one using a required idempotency key.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Portfolio account opened"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}
    },
...
    "/instruments/ovdp": {"post":{"operationId":"createManualOvdpInstrument","summary":"Create a manual ОВДП instrument","description":"Records an ОВДП instrument with absolute face value, currency, dates, and explicit coupon knowledge.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Instrument created"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/portfolio-transactions": {"post":{"operationId":"recordPortfolioTransaction","summary":"Record immutable portfolio activity","description":"Records an opening, buy, sell, coupon, redemption, or correction with position-version fencing.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Portfolio transaction posted"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
    "/portfolio-transactions/{id}/reversals": {"post":{"operationId":"reversePortfolioTransaction","summary":"Reverse portfolio activity","description":"Appends an exact compensating fact and coordinates cancellation or reversal of optional cash settlement.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"201":{"description":"Portfolio reversal posted"},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"$ref":"#/components/responses/Conflict"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},
...
    "/subscriptions/{subscription_id}": {
      "get": {"operationId":"getSubscription","summary":"Get a subscription","description":"Returns one recurring subscription and its current optimistic version.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Subscription"},"500":{"$ref":"#/components/responses/InternalServerError"}}},
      "patch": {"operationId":"updateSubscription","summary":"Update a subscription","description":"Changes subscription lifecycle or forecast metadata using expected_version.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Updated subscription"},"500":{"$ref":"#/components/responses/InternalServerError"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
      "IdempotencyKey": {"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":1,"maxLength":200}},
...
    },
    "responses": {
      "BadRequest": {"description": "Invalid request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Unauthorized": {"description": "Authentication required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "Conflict": {"description": "Optimistic version or uniqueness conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "InternalServerError": {"description": "Supporting context unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {"error": {"type": "string"}}
      }
    }
  }
}